+++ CHANGELOG.md
@@ -1,13 +1,13 @@
-# CommonOPF Changelog
+# CommonOPF Change log
 
-## dev on main branch
+## v0.4.4
 - add `Generator <: AbstractBus` struct for P-V busses in BusInjectionModel.jl
 - add `Yij_per_unit` and `Yij` admittance matrix getters
 - add `sj(j, net)`, and `sj_per_unit` power getters
 - add `connected_busses(n, net)`
 - make it so `Network[(bus1, bus2)]` _and_ `Network[(bus2, bus1)]` return the edge struct
 
-## v0.4.3 
+## v0.4.3
 - fix OpenDSSDirect.jl at v0.9.9 because v0.9.8 has some issues
 
 ## v0.4.2